requirements.txt
# API clients
requests>=2.31.0
openai>=1.50.0
massive>=1.0.0
finnhub-python>=2.4.0
websocket-client>=1.6.0

# Data validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Logging
loguru==0.7.2

# Scheduling
schedule==1.2.1

# Utilities
python-dateutil==2.8.2
tenacity>=8.2.0  # Retry logic with exponential backoff

# Testing
pytest==8.0.0
pytest-mock==3.12.0

# Linting & Formatting
ruff>=0.8.0
